\name{aRTconn-showDbs}

\alias{showDbs}
\alias{showDbs,aRTconn-method}

\title{Show available databases}
\description{
Function that lists all databases of a given DBMS. This function
lists only the databases the user has permission to read.
}

\usage{
\S4method{showDbs}{aRTconn}(object)
}

\arguments{
  \item{object}{object of class \code{aRTconn}.}
}

\author{Marcos Aurelio Carrero, Pedro Ribeiro de Andrade}
\keyword{database}
\seealso{
	\code{aRTconn-class} shows the description of this class.
}
% vim: syntax=tex
