\docType{methods}

\alias{openLayer}
\alias{openLayer,aRTdb-method}

\title{Open a Layer}

\name{aRTdb-openLayer}

\description{Opens a database layer.}

\usage{
\S4method{openLayer}{aRTdb}(object, lname)
}

\arguments{
   \item{object}{Object of aRTdb class.}
   \item{lname}{Name of the layer.}
}

\author{Pedro Ribeiro de Andrade}
\keyword{database}
\seealso{
  See \code{\link{aRTlayer-class}} to the description of this class, and
  \code{\link{createLayer}} for creating layers.
}
% vim: syntax=tex
