\name{aRTdb-showLayers}
\alias{showLayers}
\alias{showLayers,aRTdb-method}
\title{Show the available layers}
\description{
	Function that shows all available layers in the database.
}
\usage{
\S4method{showLayers}{aRTdb}(object)
}
\arguments{
  \item{object}{Object of class \code{aRTdb}.}
}

\author{Marcos Aurelio Carrero, Pedro Ribeiro de Andrade}
\keyword{database}
\seealso{
	See \code{aRTdb-class} to show the description of this class.
}
% vim: syntax=tex
