\name{aRTdb-showThemes}
\alias{showThemes}
\alias{showThemes,aRTdb-method}
\title{showThemes}
\description{
	Function that shows all themes in the database.
}
\usage{
\S4method{showThemes}{aRTdb}(object)
}
\arguments{
  \item{object}{Object of class \code{aRTdb}.}
}

\author{Marcos Aurelio Carrero, Pedro Ribeiro de Andrade}
\keyword{database}
\seealso{
	See \code{aRTdb-class} to show the description of this class.
}
% vim: syntax=tex
