\name{aRTdb-showViews}
\alias{showViews}
\alias{showViews,aRTdb-method}
\title{Show database views}
\description{
	Function that shows all views of the database.
}
\usage{
\S4method{showViews}{aRTdb}(object)
}
\arguments{
  \item{object}{Object of aRTdb class.}
}

\author{Marcos Aurelio Carrero, Pedro Ribeiro de Andrade}
\keyword{database}
\seealso{
	See \code{aRTdb-class} to show the description of this class.
}
% vim: syntax=tex
