\name{aRTgeometry}
\alias{aRTgeometry}
\alias{aRTgeometry-class}
\alias{initialize,aRTgeometry-method}

% TODO: tirar isso daqui e colocar na sua propria documentacao
\alias{removeDup}
\alias{removeDup,aRTgeometry-method}

\title{aRTgeometry}

\description{
	An aRTgeometry object represents a TerraLib geometry in memory. Using it, one
    can execute geometric operations without storing the data in a database. 

	Objects of class aRTgeometry can be directly inserted into layers, without
    converting them to Spatial objects.
	
	aRTgeometry supports the three geometric types: points, polygons and lines.
}

\author{Pedro Ribeiro de Andrade}
\keyword{database}
% vim: syntax=tex
