\name{getGeometry}
\alias{getGeometry,aRTgeometry-method}
\alias{getGeometry}
\title{getGeometry}

\description{
Returns the conversion of an \pkg{aRTgeometry} object to the Spatial format.
}

\usage{
\S4method{getGeometry}{aRTgeometry}(object, calcal=TRUE)
}

\arguments{
  \item{object}{Object of \pkg{aRTgeometry} class.}
  \item{calcal}{CALCulate Area and Labpt. Otherwise these values will be empty.}
}

\author{Pedro Ribeiro de Andrade}
\keyword{database}

% vim: syntax=tex
