\name{getID}
\alias{getID}
\title{getID}
\alias{getID,SpatialPointsDataFrame-method}
\alias{getID,SpatialLines-method}
\alias{getID,SpatialPolygons-method}
\alias{getID,aRTlayer-method}

\description{
Returns the IDs of a Spatial object or from an aRTlayer.
}

\usage{
getID(object)
}

\arguments{
  \item{object}{Object of a \pkg{sp} class.}
}

\author{Pedro Ribeiro de Andrade}
\keyword{database}

% vim: syntax=tex
