\name{aRTlayer-addPoints}
\alias{addPoints}
\alias{addPoints,aRTlayer-method}
\title{Add points to a layer}

\description{
	Function to add a set of points to a layer.
}
\usage{
\S4method{addPoints}{aRTlayer}(object, points)
}
\arguments{
  \item{object}{Object of class \code{aRTlayer}.}
  \item{points}{A set of points described in an \code{aRTgeometry} object.}
}

\author{Marcos Aurelio Carrero, Pedro Ribeiro de Andrade}
\keyword{connection}
\seealso{
	See \code{aRTlayer-class} to show the description of this class.
}
% vim: syntax=tex
