\name{aRTlayer-addPolygons}

\alias{addPolygons}
\alias{addPolygons,aRTlayer-method}

\title{Add polygons to a layer}

\description{
	Function to add a set of polygons to a layer.
}

\usage{
\S4method{addPolygons}{aRTlayer}(object, polygons)
}

\arguments{
  \item{object}{Object of class \code{aRTlayer}.}
  \item{polygons}{A Set of polygons }
}

\author{Marcos Aurelio Carrero, Pedro Ribeiro de Andrade}
\keyword{connection}
\seealso{
	See \code{aRTlayer-class} to show the description of this class.
}
% vim: syntax=tex
