\name{aRTlayer-addRaster}

\alias{addRaster}
\alias{addRaster,aRTlayer-method}

\title{Add a raster data to a layer}

\description{
	Function to add a raster (image) to a layer.
}

\usage{
\S4method{addRaster}{aRTlayer}(object, raster)
}

\arguments{
  \item{object}{Object of class \code{aRTlayer}.}
  \item{raster}{A raster data in a matrix format.}
}

\author{Marcos Aurelio Carrero, Pedro Ribeiro de Andrade}
\keyword{connection}
\seealso{
	See \code{aRTlayer-class} to show the description of this class.
}
% vim: syntax=tex
