\name{aRTlayer-getBox}
\alias{getBox}
\alias{getBox,aRTlayer-method}
\title{Bounding box}

\description{
	Function that returns the bounding box of a given layer.
}

\usage{
\S4method{getBox}{aRTlayer}(object)
}

\arguments{
  \item{object}{Object of class \code{aRTlayer}.}
}

\author{Marcos Aurelio Carrero, Pedro Ribeiro de Andrade}
\keyword{connection}
\seealso{
   See \code{aRTlayer-class} to the description of this class.
}
% vim: syntax=tex
