\name{aRTlayer-getProj}
\alias{getProj}
\alias{getProj,aRTlayer-method}
\title{Layer's projection}

\description{
	Function that returns the layer's projection, following the proj.4 format.
}

\usage{
\S4method{getProj}{aRTlayer}(object)
}

\arguments{
  \item{object}{Object of class \code{aRTlayer}.}
}

\author{Pedro Ribeiro de Andrade}
\keyword{connection}
\seealso{
   See \code{aRTlayer-class} to the description of this class.
}
% vim: syntax=tex
