\name{aRTlayer-getRaster}

\alias{getRaster}
\alias{getRaster,aRTlayer-method}

\title{Raster data from a layer}

\description{
	Function that returns the raster of a layer.
}

\usage{
\S4method{getRaster}{aRTlayer}(object,as.sp)
}

\arguments{
  \item{object}{Object of class \code{aRTlayer}.}
  \item{as.sp}{Return a Spatial data (TRUE, default) or a matrix (FALSE)?}
}

\author{Marcos Aurelio Carrero, Pedro Ribeiro de Andrade}
\keyword{connection}
\seealso{
   See \code{aRTlayer-class} to the description of this class.
}
% vim: syntax=tex
