\name{aRTnetwork}

\docType{class}

\alias{aRTnetwork}
\alias{aRTnetwork-class}
\alias{initialize,aRTnetwork-method}
\alias{createNetwork}

\title{Class aRTnetwork}
\description{Class that represents a set of lines composing a network, for calculating non-Euclidean distances.}

\section{Objects from the Class}{
Objects can be created by using \code{createNetwork(lines, weight)}.
}

\section{Arguments}{
   \describe{
   \item{lines}{A SpatialLines object.}
   \item{weight}{The weight of each line.}
   }
}
				 
\section{Extends}{
Class \code{aRTcomponent}, directly.
}

\author{Pedro Ribeiro de Andrade}
\keyword{classes}
\seealso{
 See \code{aRTnetwork-METHOD} for a method of this class or \code{CLASSNAME-class} for links to other classes.
}

% vim: syntax=tex
