\name{aRTraster-getPixels}

\alias{getPixels}
\alias{getPixels,aRTraster-method}

\title{Select pixels within geometry}

\description{
	Function that returns the pixels which satisfies some relation with a given geometry.
}

\usage{
\S4method{getPixels}{aRTraster}(object,geometry, mode)
}

\arguments{
  \item{object}{Object of class \code{aRTraster}.}
  \item{geometry}{The geometry. Currently it works only with points and polygons.}
  \item{mode}{under development.}
}

\author{Marcos Aurelio Carrero, Pedro Ribeiro de Andrade}
\keyword{connection}
\seealso{
   See \code{aRTraster-class} to the description of this class.
}
% vim: syntax=tex
