\docType{methods}
\alias{getVisual}
\alias{getVisual,aRTtheme-method}
\alias{getVisual,aRTvisual-method}

\title{Get the visual configuration}

\name{aRTtheme-getVisual}

\description{
	TODO (not implemented)
Returns a list containing the visual configuration of the object.
}

\usage{
\S4method{getVisual}{aRTtheme}(object)
\S4method{getVisual}{aRTvisual}(object)
}

\arguments{
  \item{object}{Object of class \{\code{aRTtheme},\code{aRTvisual}\}.}
}

\author{Pedro Ribeiro de Andrade}
\keyword{database}
\seealso{
   See \code{aRTtheme-class} and \code{aRTvisual-class} for the description of these classes.
}

% vim: syntax=tex
