\name{visualPoints}
\alias{visualPoints}
\alias{visualPointsSlice}

\title{Visual points}
\description{Build an \code{aRTvisual} object using \code{plot} 
parameters.
}

\usage{
visualPoints(pch, color, size)
}

\arguments{
	\item{color}{color of the elements.}
	\item{pch}{the style of the drawing (see below).}
	\item{size}{the size of the object to be drawn.} 
}

\section{Supported point styles}{
\itemize{
	\item 19
	\item 21
	\item 22
	\item 23
	\item 26
	\item 27
	\item \dQuote{+}
    \item \dQuote{*}
    \item \dQuote{x}
}}

\author{Pedro Ribeiro de Andrade}
\keyword{misc}
\seealso{
 See \code{aRTvisual-METHOD} for a method of this class or \code{CLASSNAME-class} for links to other classes.
}

% vim: syntax=tex
