\name{visualPolygons}
\alias{visualPolygons}

\title{Visual polygons}
\description{Build an \code{aRTvisual} object using \code{plot} 
parameters.
}

\usage{
visualPolygons(...)
}

\arguments{
%	\item{color}{internal color of the elements, or a vector of colors of a raster.}
%	\item{ccolor}{the contour color.}
%	\item{style}{the style of the drawing. It depends on the geometry.}
%	\item{transp}{the transparency factor of the color, a integer between 0 and 100.}
	\item{...}{Other arguments of an aRTvisual object.}
%	\item{ctransp}{the transparency factor of ccolor.}
%	\item{cwidth}{the contour width.}
%	\item{width}{width of the line.}
%	\item{size}{the size of the object to be drawn.} 
}

\section{Polygon styles}{
\describe{
  \item{0}{aa}
  \item{1}{aa}
}}


\author{Pedro Ribeiro de Andrade}
\keyword{misc}
\seealso{
 See \code{aRTvisual-METHOD} for a method of this class or \code{CLASSNAME-class} for links to other classes.
}

% vim: syntax=tex
