\name{visualRaster}
\alias{visualRaster}

\title{Visual raster}
\description{Build an \code{aRTvisual} object using \code{plot} 
parameters.
}

\usage{
visualRaster(color = terrain.colors(10))
}

\arguments{
	\item{color}{vector of colors.}
}

\author{Pedro Ribeiro de Andrade}
\keyword{misc}
\seealso{
 See \code{aRTvisual-METHOD} for a method of this class or \code{CLASSNAME-class} for links to other classes.
}

% vim: syntax=tex
