\name{GPM}

\alias{createGPM,aRTlayer-method}
\alias{createGPM}
\alias{addEntryPoints}
\alias{calculateDistancePoints}
\alias{connectToBiggerIntersectionArea}
\alias{connectLineToIntersectionPolygons}
\alias{createOpenNetwork}
\alias{getNeighborsContainedPoints}
\alias{getNeighborsIntersectionLines}
\alias{getNeighborsDistanceFunction}
\alias{getNeighborsMaxDistanceFunction}
\alias{getNeighborsOpenNetworkFunction}
\alias{saveGPM}

\title{GPM}

\description{
	GPM is currently under development. Please read the vignette for more information.
}

\author{Pedro Ribeiro de Andrade}
\keyword{database}
% vim: syntax=tex
