\name{recife}
\alias{recife.points}
\alias{recife.sewers}
\alias{recife.districts}
\alias{recife}
\title{recife}
\description{
	Recife city, in Brazil. It is represented by three datasets: points,
	sewers and districts.
}

\usage{data(recife.points); data(recife.sewers); data(recife.districts)}

\author{Pedro Ribeiro de Andrade}
\keyword{datasets}
% vim: syntax=tex
