\name{aRTsilent}

\alias{aRTsilent}

\title{Silent mode}

\description{
	Set aRT silent mode.
}

\usage{
aRTsilent(value)
}

\arguments{
  \item{value}{Silent mode? If \code{NULL}, this function returns if aRT is in silent mode.}
}

\author{Pedro Ribeiro de Andrade}
\keyword{misc}

\seealso{
	See \code{aRTCLASS-class} to show the description of CLASS.
}
% vim: syntax=tex
