\alias{toDate}
\name{toDate}

\title{Convert dates}
\description{Convert a data in numeric format (year, month, etc.) to a string to be used as aRT temporal data.}
\usage{
	toDate(year = 0, month = 1, day = 1, hour = 0, minute = 0, second = 0)
}

\arguments{
   \item{year}{The year.}
   \item{month}{The month (1-12).}
   \item{day}{The day of the month.}
   \item{hour}{The hour (0-23)}
   \item{minute}{The minute.}
   \item{second}{The second.}
}

\author{Pedro Ribeiro de Andrade}
\keyword{chron}

% vim: syntax=tex
