\name{aRTversion}

\docType{methods}

\title{Version of aRT package}

\alias{aRTversion}
\description{Return a string containing the package's version.}
\usage{
	aRTversion()
}
\author{Marcos Aurelio Carrero, Pedro Ribeiro de Andrade}

\keyword{misc}

\seealso{
   See \code{\link{aRT-classes}} for helping about the package organization.
}

% vim: syntax=tex
